Big dump, refactorings for fmi3
