+++ fmi-schema/src/fmi3/type.rs
@@ -2,7 +2,12 @@ use yaserde_derive::{YaDeserialize, YaSerialize};
 
 use super::{Float32Attributes, Float64Attributes, RealBaseAttributes};
 
-#[derive(Clone, Default, PartialEq, Debug, YaSerialize, YaDeserialize)]
+pub trait BaseTypeTrait {
+    fn name(&self) -> &str;
+    fn description(&self) -> Option<&str>;
+}
+
+#[derive(Default, PartialEq, Debug, YaSerialize, YaDeserialize)]
 pub struct TypeDefinitionBase {
     #[yaserde(attribute)]
     pub name: String,
@@ -10,7 +15,7 @@ pub struct TypeDefinitionBase {
     pub description: Option<String>,
 }
 
-#[derive(Clone, Default, PartialEq, Debug, YaSerialize, YaDeserialize)]
+#[derive(Default, PartialEq, Debug, YaSerialize, YaDeserialize)]
 pub struct Float32Type {
     #[yaserde(flatten)]
     pub base: TypeDefinitionBase,
@@ -20,7 +25,7 @@ pub struct Float32Type {
     pub attr: Float32Attributes,
 }
 
-#[derive(Clone, Default, PartialEq, Debug, YaSerialize, YaDeserialize)]
+#[derive(Default, PartialEq, Debug, YaSerialize, YaDeserialize)]
 pub struct Float64Type {
     #[yaserde(flatten)]
     pub base: TypeDefinitionBase,
@@ -28,4 +33,24 @@ pub struct Float64Type {
     pub base_attr: RealBaseAttributes,
     #[yaserde(flatten)]
     pub attr: Float64Attributes,
+}
+
+impl BaseTypeTrait for Float32Type {
+    fn name(&self) -> &str {
+        &self.base.name
+    }
+
+    fn description(&self) -> Option<&str> {
+        self.base.description.as_deref()
+    }
+}
+
+impl BaseTypeTrait for Float64Type {
+    fn name(&self) -> &str {
+        &self.base.name
+    }
+
+    fn description(&self) -> Option<&str> {
+        self.base.description.as_deref()
+    }
 }
