+++ src/fmi2/instance/model_exchange.rs
@@ -1,7 +1,5 @@
 use std::ffi::CString;
 
-use libc::bind;
-
 use crate::{
     fmi2::{
         binding::{self},
@@ -13,12 +11,12 @@ use crate::{
     FmiError, FmiResult,
 };
 
-use super::{traits, Instance, ME};
+use super::{Instance, ME};
 
 impl<'a> Instance<'a, ME> {
     /// Initialize a new Instance from an Import
     pub fn new(
-        import: &import::Fmi2,
+        import: &'a import::Fmi2,
         instance_name: &str,
         visible: bool,
         logging_on: bool,
@@ -36,12 +34,13 @@ impl<'a> Instance<'a, ME> {
             CString::new(import.resource_url().as_str()).expect("Error building CString");
 
         let component = unsafe {
+            let callback_functions = &*callbacks as *const CallbackFunctions;
             binding.fmi2Instantiate(
                 instance_name.as_ptr(),
                 binding::fmi2Type_fmi2ModelExchange,
                 guid.as_ptr(),                      // guid
                 resource_url.as_ptr(),              // fmu_resource_location
-                &*callbacks,                        // functions
+                callback_functions as _,            // functions
                 visible as binding::fmi2Boolean,    // visible
                 logging_on as binding::fmi2Boolean, // logging_on
             )
@@ -73,109 +72,107 @@ impl<'a> Instance<'a, ME> {
             next_event_time: 0.0,
         };
 
-        while (event_info.new_discrete_states_needed as _ == true)
-            && (event_info.terminate_simulation as _ == true)
+        while (event_info.new_discrete_states_needed != 0) && (event_info.terminate_simulation != 0)
         {
             log::trace!("Iterating while new_discrete_states_needed=true");
-            self.new_discrete_states(&mut event_info)?;
+            //self.new_discrete_states(&mut event_info)?;
+            todo!();
         }
 
-        assert_eq!(
-            event_info.terminate_simulation as _, false,
+        assert!(
+            event_info.terminate_simulation != 0,
             "terminate_simulation in=true do_event_iteration!"
         );
 
         Ok((
-            event_info.nominals_of_continuous_states_changed as _ == true,
-            event_info.values_of_continuous_states_changed as _ == true,
+            event_info.nominals_of_continuous_states_changed != 0,
+            event_info.values_of_continuous_states_changed != 0,
         ))
     }
 }
 
 impl<'a> ModelExchange for Instance<'a, ME> {
-    fn enter_event_mode(&self) -> FmiResult<FmiStatus> {
-        unsafe { self.binding.fmi2EnterEventMode(self.component) }.into()
+    fn enter_event_mode(&self) -> FmiStatus {
+        FmiStatus(unsafe { self.binding.fmi2EnterEventMode(self.component) })
     }
 
-    fn new_discrete_states(&self, event_info: &mut EventInfo) -> FmiResult<FmiStatus> {
-        unsafe {
+    fn new_discrete_states(&self, event_info: &mut EventInfo) -> FmiStatus {
+        let event_info: *mut binding::fmi2EventInfo = &*event_info as *const _ as _;
+        FmiStatus(unsafe {
             self.binding
                 .fmi2NewDiscreteStates(self.component, event_info)
-        }
-        .into()
+        })
     }
 
-    fn enter_continuous_time_mode(&self) -> FmiResult<FmiStatus> {
-        unsafe { self.binding.fmi2EnterContinuousTimeMode(self.component) }.into()
+    fn enter_continuous_time_mode(&self) -> FmiStatus {
+        FmiStatus(unsafe { self.binding.fmi2EnterContinuousTimeMode(self.component) })
     }
 
     fn completed_integrator_step(
         &self,
         no_set_fmu_state_prior_to_current_point: bool,
-    ) -> FmiResult<(bool, bool)> {
+    ) -> (FmiStatus, bool, bool) {
         // The returned tuple are the flags (enter_event_mode, terminate_simulation)
-        let mut enter_event_mode = false as _;
-        let mut terminate_simulation = false as _;
-        let res: FmiResult<FmiStatus> = unsafe {
+        let mut enter_event_mode = false;
+        let mut terminate_simulation = false;
+        let res = FmiStatus(unsafe {
             self.binding.fmi2CompletedIntegratorStep(
                 self.component,
                 no_set_fmu_state_prior_to_current_point as binding::fmi2Boolean,
-                &mut enter_event_mode,
-                &mut terminate_simulation,
+                &mut (enter_event_mode as _),
+                &mut (terminate_simulation as _),
             )
-        }
-        .into();
-        res.and(Ok((
-            enter_event_mode as _ == true,
-            terminate_simulation as _ == true,
-        )))
+        });
+        (res, enter_event_mode, terminate_simulation)
     }
 
-    fn set_time(&self, time: f64) -> FmiResult<FmiStatus> {
-        let res: FmiStatus = unsafe {
+    fn set_time(&self, time: f64) -> FmiStatus {
+        FmiStatus(unsafe {
             self.binding
                 .fmi2SetTime(self.component, time as binding::fmi2Real)
-        }
-        .into();
+        })
     }
 
-    fn set_continuous_states(&self, states: &[f64]) -> FmiResult<FmiStatus> {
+    fn set_continuous_states(&self, states: &[f64]) -> FmiStatus {
         assert!(states.len() == self.schema.num_states());
-        unsafe {
+        FmiStatus(unsafe {
             self.binding
                 .fmi2SetContinuousStates(self.component, states.as_ptr(), states.len())
-        }
-        .into()
+        })
     }
 
-    fn get_derivatives(&self, dx: &mut [f64]) -> FmiResult<FmiStatus> {
+    fn get_derivatives(&self, dx: &mut [f64]) -> FmiStatus {
         assert!(dx.len() == self.schema.num_states());
-        unsafe {
+        FmiStatus(unsafe {
             self.binding
                 .fmi2GetDerivatives(self.component, dx.as_mut_ptr(), dx.len())
-        }
-        .into()
+        })
     }
 
-    fn get_event_indicators(&self, events: &mut [f64]) -> FmiResult<FmiStatus> {
-        assert!(events.len() == self.schema.num_event_indicators());
-        unsafe {
+    fn get_event_indicators(&self, events: &mut [f64]) -> FmiStatus {
+        assert_eq!(events.len(), self.schema.num_event_indicators());
+        FmiStatus(unsafe {
             self.binding
                 .fmi2GetEventIndicators(self.component, events.as_mut_ptr(), events.len())
-        }
-        .into()
+        })
     }
 
-    fn get_continuous_states(&self, states: &mut [f64]) -> FmiResult<FmiStatus> {
-        assert!(states.len() == self.scham.num_states());
-        unsafe {
+    fn get_continuous_states(&self, states: &mut [f64]) -> FmiStatus {
+        assert_eq!(states.len(), self.schema.num_states());
+        FmiStatus(unsafe {
             self.binding
                 .fmi2GetContinuousStates(self.component, states.as_mut_ptr(), states.len())
-        }
-        .into()
+        })
     }
 
-    fn get_nominals_of_continuous_states(&self) -> FmiResult<&[f64]> {
-        unimplemented!();
+    fn get_nominals_of_continuous_states(&self, nominals: &mut [f64]) -> FmiStatus {
+        assert_eq!(nominals.len(), self.schema.num_states());
+        FmiStatus(unsafe {
+            self.binding.fmi2GetNominalsOfContinuousStates(
+                self.component,
+                nominals.as_mut_ptr() as _,
+                nominals.len(),
+            )
+        })
     }
 }
