+++ src/fmi2/instance/traits.rs
@@ -1,9 +1,6 @@
-use crate::{
-    fmi2::{binding, meta, EventInfo},
-    FmiResult,
-};
+//! Traits for different instance types ([ModelExchange], [CoSimulation]).
 
-use super::FmiStatus;
+use super::{binding, EventInfo, FmiStatus, StatusKind};
 
 /// Interface common to both ModelExchange and CoSimulation
 pub trait Common {
@@ -11,9 +8,9 @@ pub trait Common {
     fn name(&self) -> &str;
 
     /// The FMI-standard version string
-    fn version(&self) -> FmiResult<&str>;
+    fn version(&self) -> &str;
 
-    fn set_debug_logging(&self, logging_on: bool, categories: &[&str]) -> FmiResult<()>;
+    fn set_debug_logging(&self, logging_on: bool, categories: &[&str]) -> FmiStatus;
 
     /// Informs the FMU to setup the experiment. This function can be called after `instantiate()`
     /// and before `enter_initialization_mode()` is called.
@@ -44,7 +41,7 @@ pub trait Common {
         tolerance: Option<f64>,
         start_time: f64,
         stop_time: Option<f64>,
-    ) -> FmiResult<FmiStatus>;
+    ) -> FmiStatus;
 
     /// Informs the FMU to enter Initialization Mode.
     ///
@@ -53,33 +50,49 @@ pub trait Common {
     /// *Setting other variables is not allowed*. Furthermore, `setup_experiment()` must be called
     /// at least once before calling `enter_initialization_mode()`, in order that `start_time` is
     /// defined.
-    fn enter_initialization_mode(&self) -> FmiResult<FmiStatus>;
+    fn enter_initialization_mode(&self) -> FmiStatus;
 
     /// Informs the FMU to exit Initialization Mode.
     ///
     /// Under ModelExchange this function switches off all initialization equations and the FMU
     /// enters implicitely Event Mode, that is all continuous-time and active discrete-time
     /// equations are available.
-    fn exit_initialization_mode(&self) -> FmiResult<FmiStatus>;
+    fn exit_initialization_mode(&self) -> FmiStatus;
 
     /// Informs the FMU that the simulation run is terminated.
     ///
     /// After calling this function, the final values of all variables can be inquired with the
     /// fmi2GetXXX(..) functions. It is not allowed to call this function after one of the
     /// functions returned with a status flag of fmi2Error or fmi2Fatal.
-    fn terminate(&self) -> FmiResult<FmiStatus>;
+    fn terminate(&self) -> FmiStatus;
 
     /// Is called by the environment to reset the FMU after a simulation run.
     ///
     /// The FMU goes into the same state as if fmi2Instantiate would have been called. All
     /// variables have their default values. Before starting a new run, fmi2SetupExperiment and
     /// fmi2EnterInitializationMode have to be called.
-    fn reset(&self) -> FmiResult<FmiStatus>;
+    fn reset(&self) -> FmiStatus;
 
-    fn get_real(&self, sv: &meta::ScalarVariable) -> FmiResult<binding::fmi2Real>;
-    fn get_integer(&self, sv: &meta::ScalarVariable) -> FmiResult<binding::fmi2Integer>;
-    fn get_boolean(&self, sv: &meta::ScalarVariable) -> FmiResult<binding::fmi2Boolean>;
-    fn get_string(&self, sv: &meta::ScalarVariable) -> FmiResult<binding::fmi2String>;
+    fn get_real(
+        &self,
+        sv: &[binding::fmi2ValueReference],
+        v: &mut [binding::fmi2Real],
+    ) -> FmiStatus;
+    fn get_integer(
+        &self,
+        sv: &[binding::fmi2ValueReference],
+        v: &mut [binding::fmi2Integer],
+    ) -> FmiStatus;
+    fn get_boolean(
+        &self,
+        sv: &[binding::fmi2ValueReference],
+        v: &mut [binding::fmi2Boolean],
+    ) -> FmiStatus;
+    fn get_string(
+        &self,
+        sv: &[binding::fmi2ValueReference],
+        v: &mut [binding::fmi2String],
+    ) -> FmiStatus;
 
     /// Set real values
     ///
@@ -88,9 +101,9 @@ pub trait Common {
     /// * `values` - a slice of `fmi::fmi2Real` values to set
     fn set_real(
         &self,
-        vrs: &[meta::ValueReference],
+        vrs: &[binding::fmi2ValueReference],
         values: &[binding::fmi2Real],
-    ) -> FmiResult<FmiStatus>;
+    ) -> FmiStatus;
 
     /// Set integer values
     ///
@@ -99,21 +112,21 @@ pub trait Common {
     /// * `values` - a slice of `fmi::fmi2Integer` values to set
     fn set_integer(
         &self,
-        vrs: &[meta::ValueReference],
+        vrs: &[binding::fmi2ValueReference],
         values: &[binding::fmi2Integer],
-    ) -> FmiResult<FmiStatus>;
+    ) -> FmiStatus;
 
     fn set_boolean(
         &self,
-        vrs: &[meta::ValueReference],
-        values: &[binding::fmi2Boolean],
-    ) -> FmiResult<FmiStatus>;
+        vrs: &[binding::fmi2ValueReference],
+        values: &mut [binding::fmi2Boolean],
+    ) -> FmiStatus;
 
     fn set_string(
         &self,
-        vrs: &[meta::ValueReference],
+        vrs: &[binding::fmi2ValueReference],
         values: &[binding::fmi2String],
-    ) -> FmiResult<FmiStatus>;
+    ) -> FmiStatus;
 
     // fn get_fmu_state(&self) -> Result<FmuState>;
     // fn set_fmu_state(&self, state: &FmuState<Self::Api>) -> Result<()>;
@@ -133,11 +146,11 @@ pub trait Common {
     /// used to construct the desired partial derivative matrices.
     fn get_directional_derivative(
         &self,
-        unknown_vrs: &[meta::ValueReference],
-        known_vrs: &[meta::ValueReference],
+        unknown_vrs: &[binding::fmi2ValueReference],
+        known_vrs: &[binding::fmi2ValueReference],
         dv_known_values: &[binding::fmi2Real],
         dv_unknown_values: &mut [binding::fmi2Real],
-    ) -> FmiResult<FmiStatus>;
+    ) -> FmiStatus;
 }
 
 pub trait ModelExchange: Common {
@@ -145,7 +158,7 @@ pub trait ModelExchange: Common {
 
     /// The model enters Event Mode from the Continuous-Time Mode and discrete-time equations may
     /// become active (and relations are not "frozen").
-    fn enter_event_mode(&self) -> FmiResult<FmiStatus>;
+    fn enter_event_mode(&self) -> FmiStatus;
 
     /// The FMU is in Event Mode and the super dense time is incremented by this call. If the super
     /// dense time before a call to `new_discrete_states` was (tR,tI) then the time instant after
@@ -159,7 +172,7 @@ pub trait ModelExchange: Common {
     ///     * call `enter_continuous_time_mode` if all FMUs return `new_discrete_states_needed` =
     ///       false.
     ///     * stay in Event Mode otherwise.
-    fn new_discrete_states(&self, event_info: &mut EventInfo) -> FmiResult<FmiStatus>;
+    fn new_discrete_states(&self, event_info: &mut EventInfo) -> FmiStatus;
 
     /// The model enters Continuous-Time Mode and all discrete-time equations become inactive and
     /// all relations are "frozen".
@@ -167,7 +180,7 @@ pub trait ModelExchange: Common {
     /// This function has to be called when changing from Event Mode (after the global event
     /// iteration in Event Mode over all involved FMUs and other models has converged) into
     /// Continuous-Time Mode.
-    fn enter_continuous_time_mode(&self) -> FmiResult<FmiStatus>;
+    fn enter_continuous_time_mode(&self) -> FmiStatus;
 
     /// Complete integrator step and return enterEventMode.
     ///
@@ -181,13 +194,13 @@ pub trait ModelExchange: Common {
     fn completed_integrator_step(
         &self,
         no_set_fmu_state_prior_to_current_point: bool,
-    ) -> FmiResult<(bool, bool)>;
+    ) -> (FmiStatus, bool, bool);
 
     /// Set a new time instant and re-initialize caching of variables that depend on time, provided
     /// the newly provided time value is different to the previously set time value (variables that
     /// depend solely on constants or parameters need not to be newly computed in the sequel, but
     /// the previously computed values can be reused).
-    fn set_time(&self, time: f64) -> FmiResult<FmiStatus>;
+    fn set_time(&self, time: f64) -> FmiStatus;
 
     /// Set a new (continuous) state vector and re-initialize caching of variables that depend on
     /// the states. Argument nx is the length of vector x and is provided for checking purposes
@@ -195,24 +208,24 @@ pub trait ModelExchange: Common {
     /// newly computed in the sequel, but the previously computed values can be reused).
     /// Note, the continuous states might also be changed in Event Mode.
     /// Note: fmi2Status = fmi2Discard is possible.
-    fn set_continuous_states(&self, states: &[f64]) -> FmiResult<FmiStatus>;
+    fn set_continuous_states(&self, states: &[f64]) -> FmiStatus;
 
     /// Compute state derivatives and event indicators at the current time instant and for the
     /// current states. The derivatives are returned as a vector with “nx” elements.
-    fn get_derivatives(&self, dx: &mut [f64]) -> FmiResult<FmiStatus>;
+    fn get_derivatives(&self, dx: &mut [f64]) -> FmiStatus;
 
     /// A state event is triggered when the domain of an event indicator changes from zj > 0 to zj ≤
     /// 0 or vice versa. The FMU must guarantee that at an event restart zj ≠ 0, for example by
     /// shifting zj with a small value. Furthermore, zj should be scaled in the FMU with its nominal
     /// value (so all elements of the returned vector “eventIndicators” should be in the order of
     /// “one”). The event indicators are returned as a vector with “ni” elements.
-    fn get_event_indicators(&self, events: &mut [f64]) -> FmiResult<FmiStatus>;
+    fn get_event_indicators(&self, events: &mut [f64]) -> FmiStatus;
 
     /// Return the new (continuous) state vector x.
     /// This function has to be called directly after calling function `enter_continuous_time_mode`
     /// if it returns with eventInfo->valuesOfContinuousStatesChanged = true (indicating that the
     /// (continuous-time) state vector has changed).
-    fn get_continuous_states(&self, x: &mut [f64]) -> FmiResult<FmiStatus>;
+    fn get_continuous_states(&self, x: &mut [f64]) -> FmiStatus;
 
     /// Return the nominal values of the continuous states. This function should always be called
     /// after calling function new_discrete_states if it returns with
@@ -226,7 +239,7 @@ pub trait ModelExchange: Common {
     /// Note, it is required that x_nominal[i] > 0.0 [Typically, the nominal values of the
     /// continuous states are used to compute the absolute tolerance required by the integrator.
     /// Example: absoluteTolerance[i] = 0.01*tolerance*x_nominal[i];]
-    fn get_nominals_of_continuous_states(&self) -> FmiResult<&[f64]>;
+    fn get_nominals_of_continuous_states(&self, nominals: &mut [f64]) -> FmiStatus;
 }
 
 pub trait CoSimulation: Common {
@@ -244,7 +257,7 @@ pub trait CoSimulation: Common {
         current_communication_point: f64,
         communication_step_size: f64,
         new_step: bool,
-    ) -> FmiResult<FmiStatus>;
+    ) -> FmiStatus;
 
     /// Cancel a running asynchronous step.
     ///
@@ -252,8 +265,8 @@ pub trait CoSimulation: Common {
     /// asynchronous execution. The master calls this function if e.g. the co-simulation run is
     /// stopped by the user or one of the slaves. Afterwards it is only allowed to call the
     /// functions `terminate()` or `reset()`.
-    fn cancel_step(&self) -> FmiResult<FmiStatus>;
+    fn cancel_step(&self) -> FmiStatus;
 
     /// Inquire into slave status during asynchronous step.
-    fn get_status(&self, kind: binding::fmi2StatusKind) -> FmiResult<FmiStatus>;
+    fn get_status(&self, kind: StatusKind) -> FmiStatus;
 }
