+++ src/import.rs
@@ -1,71 +1,16 @@
-use std::{io::Read, path::Path};
-
-use log::trace;
-use yaserde_derive::YaDeserialize;
+use std::{io::Read, path::Path, str::FromStr};
 
 #[cfg(feature = "fmi2")]
 use crate::fmi2;
 #[cfg(feature = "fmi3")]
 use crate::fmi3;
-use crate::{FmiError, FmiResult};
+use crate::Error;
+
+use fmi_schema::minimal::ModelDescription as MinModel;
 
 const MODEL_DESCRIPTION: &str = "modelDescription.xml";
-#[cfg(feature = "fmi2")]
-const FMI_MAJOR_VERSION_2: u32 = 2;
-#[cfg(feature = "fmi3")]
-const FMI_MAJOR_VERSION_3: u32 = 3;
 
-/// A minimal model description that only contains the FMI version
-/// This is used to determine the FMI version of the FMU
-#[derive(Debug, Default, PartialEq, YaDeserialize)]
-#[yaserde(root = "fmiModelDescription")]
-struct ModelDescription {
-    #[yaserde(attribute, rename = "fmiVersion")]
-    fmi_version: String,
-
-    #[yaserde(attribute, rename = "modelName")]
-    model_name: String,
-}
-
-impl ModelDescription {
-    /// Get the major version from the version string
-    fn major_version(&self) -> FmiResult<u32> {
-        self.fmi_version
-            .split_once(".")
-            .and_then(|(major, _)| major.parse().ok())
-            .ok_or(FmiError::Parse(format!(
-                "Unable to parse major version from FMI version string {}",
-                self.fmi_version
-            )))
-    }
-}
-
-fn extract_archive(archive: impl AsRef<Path>, outdir: impl AsRef<Path>) -> FmiResult<()> {
-    let archive = archive.as_ref();
-    let outdir = outdir.as_ref();
-    trace!("Extracting {} into {}", archive.display(), outdir.display());
-    let file = std::fs::File::open(archive)?;
-    let mut archive = zip::ZipArchive::new(file)?;
-
-    for i in 0..archive.len() {
-        let mut file = archive.by_index(i)?;
-        let outpath = outdir.join(file.name());
-        if file.is_dir() {
-            std::fs::create_dir_all(&outpath)?;
-        } else {
-            if let Some(p) = outpath.parent() {
-                if !p.exists() {
-                    std::fs::create_dir_all(p)?;
-                }
-            }
-            let mut outfile = std::fs::File::create(&outpath)?;
-            std::io::copy(&mut file, &mut outfile)?;
-        }
-    }
-    Ok(())
-}
-
-pub trait FmiImport: Sized {
+pub trait FmiImport<'a>: Sized {
     /// The raw parsed XML schema type
     type Schema;
 
@@ -73,25 +18,25 @@ pub trait FmiImport: Sized {
     type Binding;
 
     /// Create a new FMI import from a directory containing the unzipped FMU
-    fn new(dir: tempfile::TempDir, schema_xml: &str) -> FmiResult<Self>;
+    fn new(dir: tempfile::TempDir, schema_xml: String) -> Result<Self, Error>;
 
     /// Return the path to the extracted FMU
-    fn path(&self) -> &std::path::Path;
+    fn archive_path(&self) -> &std::path::Path;
 
     /// Get the path to the shared library
-    fn shared_lib_path(&self) -> FmiResult<std::path::PathBuf>;
+    fn shared_lib_path(&self) -> Result<std::path::PathBuf, Error>;
 
     /// Return the path to the resources directory
     fn resource_url(&self) -> url::Url {
-        url::Url::from_file_path(self.path().join("resources"))
+        url::Url::from_file_path(self.archive_path().join("resources"))
             .expect("Error forming resource location URL")
     }
 
     /// Get a reference to the raw-schema model description
-    fn raw_schema(&self) -> &Self::Schema;
+    fn model_description(&self) -> &Self::Schema;
 
     /// Load the plugin shared library and return the raw bindings.
-    fn raw_bindings(&self) -> FmiResult<Self::Binding>;
+    fn binding(&self) -> Result<Self::Binding, Error>;
 }
 
 /// Import is responsible for extracting the FMU, parsing the modelDescription XML and loading the
@@ -106,39 +51,37 @@ pub enum Import {
 
 impl Import {
     /// Creates a new Import by extracting the FMU and parsing the modelDescription XML
-    pub fn new(path: impl Into<std::path::PathBuf>) -> FmiResult<Import> {
-        // First create a temp directory
+    pub fn new(path: impl AsRef<Path>) -> Result<Self, Error> {
+        let file = std::fs::File::open(path)?;
+        let mut archive = zip::ZipArchive::new(file)?;
         let temp_dir = tempfile::Builder::new().prefix("fmi-rs").tempdir()?;
-        extract_archive(path.into(), temp_dir.path())?;
+        log::trace!("Extracting {archive:?} into {temp_dir:?}");
+        archive.extract(&temp_dir)?;
 
         // Open and read the modelDescription XML into a string
         let descr_file_path = temp_dir.path().join(MODEL_DESCRIPTION);
         let mut descr_file = std::fs::File::open(descr_file_path)?;
         let mut descr_buf = String::new();
         let descr_size = descr_file.read_to_string(&mut descr_buf)?;
-        trace!("Read {} bytes from {MODEL_DESCRIPTION}", descr_size,);
+        log::trace!("Read {descr_size} bytes from {MODEL_DESCRIPTION}");
 
         // Initial non-version-specific model description
-        let descr: ModelDescription =
-            yaserde::de::from_str(&descr_buf).map_err(|err| FmiError::Parse(err))?;
-
-        trace!(
+        let descr = MinModel::from_str(&descr_buf)?;
+        log::trace!(
             "Found FMI {} named '{}",
             descr.fmi_version,
             descr.model_name
         );
 
-        match descr.major_version()? {
+        match descr.version()?.major {
             #[cfg(feature = "fmi2")]
-            FMI_MAJOR_VERSION_2 => {
-                fmi2::import::Fmi2::new(temp_dir, &descr_buf).map(|import| Import::Fmi2(import))
-            }
+            2 => fmi2::import::Fmi2::new(temp_dir, &descr_buf).map(|import| Import::Fmi2(import)),
+
             #[cfg(feature = "fmi3")]
-            FMI_MAJOR_VERSION_3 => {
-                fmi3::import::Fmi3::new(temp_dir, &descr_buf).map(|import| Import::Fmi3(import))
-            }
+            3 => Ok(Self::Fmi3(fmi3::import::Fmi3::new(temp_dir, descr_buf)?)),
+
             _ => {
-                return Err(FmiError::UnsupportedFmiVersion(descr.fmi_version));
+                return Err(Error::UnsupportedFmiVersion(descr.fmi_version.to_string()));
             }
         }
     }
@@ -195,9 +138,9 @@ mod tests {
             .unwrap()
             .as_fmi3()
             .unwrap();
-        assert_eq!(import.model().fmi_version, "3.0");
-        assert_eq!(import.model().model_name, "BouncingBall");
-        let binding = import.raw_bindings().unwrap();
+        assert_eq!(import.model_description().fmi_version, "3.0");
+        assert_eq!(import.model_description().model_name, "BouncingBall");
+        let binding = import.binding().unwrap();
         let ver = unsafe {
             std::ffi::CStr::from_ptr(binding.fmi3GetVersion())
                 .to_str()
@@ -206,9 +149,7 @@ mod tests {
         assert_eq!(ver, "3.0");
 
         let mut inst1 = import.instantiate_me("inst1", true, true).unwrap();
-        inst1
-            .set_debug_logging(true, import.model().log_categories.keys())
-            .unwrap();
+        //inst1 .set_debug_logging(true, import.model().log_categories.keys()) .unwrap();
         inst1.enter_initialization_mode(None, 0.0, None).unwrap();
         inst1.exit_initialization_mode().unwrap();
         inst1.set_time(1234.0).unwrap();
@@ -216,9 +157,9 @@ mod tests {
         inst1.enter_continuous_time_mode().unwrap();
 
         let states = (0..import
-            .model()
+            .model_description()
             .model_structure
-            .continuous_state_derivatives
+            .continuous_state_derivative
             .len())
             .map(|x| x as f64)
             .collect::<Vec<_>>();
